[add] menambahkan dan menyelesaikan carousel

+++ app/services/staffService.js
@@ -1,6 +1,9 @@
 const ApplicationError = require('../../config/errors/ApplicationError');
 const staffRepository = require('../repositories/staffRepository');
 
+const fs = require('fs');
+const path = require('path');
+
 exports.getStaffs = async () => {
     try {
         const data = await staffRepository.getStaffs();
@@ -28,29 +31,22 @@ exports.createStaff = async (payload) => {
     }
 }
 
-const fs = require('fs');
-const path = require('path');
-
 exports.updateStaff = async (id, payload) => {
     try {
-        // Ambil data staff berdasarkan ID
         const staff = await staffRepository.getStaffById(id);
         if (!staff) {
             throw new ApplicationError(`Staff with id ${id} not found`, 404);
         }
 
-        // Jika image baru ada, hapus file lama
         if (payload.image) {
             const oldImagePath = path.join(__dirname, '../../uploads/staff', staff.image);
 
-            // Cek apakah file lama ada di folder
             if (fs.existsSync(oldImagePath)) {
                 fs.unlinkSync(oldImagePath); // Hapus file lama
                 console.log(`Old image deleted: ${oldImagePath}`);
             }
         }
 
-        // Update data staff dengan data baru (atau gunakan data lama jika tidak ada di payload)
         const updated = await staff.update({
             name: payload.name || staff.name,
             position: payload.position || staff.position,
@@ -66,6 +62,19 @@ exports.updateStaff = async (id, payload) => {
 
 exports.deleteStaff = async (id) => {
     try {
+        const staff = await staffRepository.getStaffById(id);
+        console.log(staff);
+        
+        if (!staff) {
+            throw new ApplicationError(`Staff with id ${id} not found`, 404);
+        }
+
+        const oldImagePath = path.join(__dirname, '../../uploads/staff', staff.image);
+
+        if (fs.existsSync(oldImagePath)) {
+            fs.unlinkSync(oldImagePath); // Hapus file lama
+            console.log(`Old image deleted: ${oldImagePath}`);
+        }
         const data = await staffRepository.deleteStaff(id);
         return data;
     } catch (error) {
